Use `tsc` to type check JS
