+++ packages/node-bridge/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "checkJs": true,
+    "lib": ["esnext"],
+    "noEmit": true,
+    "noImplicitReturns": true,
+    "strict": true,
+    "declaration": true
+  },
+  "include": ["bridge.js", "launcher.js"],
+  "exclude": ["node_modules"]
+}
